ChangeGNSolution

+++ GausNewtonSolver.hpp
@@ -10,7 +10,7 @@ public:
     void calculate_dr_db(ModelValue& current_value);
 
 
-    vector<double> Gauss_Newton(vector<double> &x0, Matrix& A, Matrix& R);
+    vector<long double> Gauss_Newton(vector<long double> &x0, Matrix& A, Matrix& R);
 
     Matrix solve_system(Matrix& gradient_f, Matrix& f_b);
 };
