+++ Structures/SimulationVector.cpp
@@ -1,6 +1,6 @@
 #include "SimulationVector.hpp"
 
-void SimulationVector::setX_vector(const std::vector<double>& newX_vector)
+void SimulationVector::setX_vector(const std::vector<long double>& newX_vector)
 {
     X_vector = newX_vector;
 }
@@ -11,17 +11,17 @@ void SimulationVector::clearX_vector()
     X_vector.resize(SIZE_VECTOR * 2);
 }
 
-void SimulationVector::setElementX_vector(int index, double value)
+void SimulationVector::setElementX_vector(int index, long double value)
 {
     X_vector[index] = value;
 }
 
-void SimulationVector::setElementDX_matrix(int y, int x, double value)
+void SimulationVector::setElementDX_matrix(int y, int x, long double value)
 {
     dX__dr0_dv0_dM.setElement(y, x, value);
 }
 
-std::vector<double> SimulationVector::getX_vector() const
+std::vector<long double> SimulationVector::getX_vector() const
 {
     return X_vector;
 }
